feat: :recycle: Separate api calls from component
