+++ src/ProfilePage/Api/profileApi.ts
@@ -7,6 +7,7 @@ import {
   CommitItem,
   LanguageData,
   Parent,
+  ProfileItem,
   RepoItem,
 } from "../../Models/interfaces";
 
@@ -34,7 +35,6 @@ export const getCommits = async (
         );
         setCommits(newCommits);
         sendUpCommits(newCommits);
-        console.log(newCommits);
         return newCommits;
       } else {
         console.error("Request failed with status:", res.status);
@@ -117,6 +117,144 @@ export const getBranches = async (
       console.error("Error fetching repositories:", error);
       setBranches([]);
       return [];
+    }
+  }
+};
+
+export const getOrgs = async (
+  url: string | undefined,
+  setOrgs: {
+    (value: SetStateAction<number | undefined>): void;
+  }
+) => {
+  if (url) {
+    try {
+      const res = await octokit.request(`GET ${url}`);
+
+      if (res.status === 200) {
+        let organizations = res.data.length;
+        if (organizations != 0) {
+          organizations -= 1;
+        }
+        setOrgs(organizations);
+        return organizations;
+      } else {
+        console.error("Request failed with status:", res.status);
+        return;
+      }
+    } catch (error) {
+      console.error("Error occurred:", error);
+      return;
+    }
+  }
+};
+
+export const getStats = async (
+  username: string | undefined,
+  setStars: { (branchNumber: number): void }
+) => {
+  if (username) {
+    try {
+      const res = await octokit.request(
+        `GET https://api.github.com/users/${username}/starred`
+      );
+
+      if (res.status === 200) {
+        let stars = res.data.length;
+        if (stars != 0) {
+          stars -= 1;
+        }
+        setStars(stars);
+        return stars;
+      } else {
+        console.error("Request failed with status:", res.status);
+        return;
+      }
+    } catch (error) {
+      console.error("Error occurred:", error);
+      return;
+    }
+  }
+};
+
+export const getRepos = async (
+  profileName: string,
+  setRepos: {
+    (value: SetStateAction<RepoItem[]>): void;
+  },
+  setRepoNumber: { (branchNumber: number): void }
+) => {
+  if (profileName) {
+    try {
+      const res = await octokit.request(
+        `GET https://api.github.com/users/${profileName}/repos?sort=updated`
+      );
+
+      if (res.status === 200) {
+        let repos = res.data.length;
+        const data = res.data;
+        const newRepos = data.map((repo: RepoItem) => ({
+          id: repo.id,
+          name: repo.name,
+          full_name: repo.full_name,
+          private: repo.private,
+          description: repo.description,
+          collaborators_url: repo.collaborators_url,
+          branches_url: repo.branches_url,
+          contributors_url: repo.contributors_url,
+          commits_url: repo.commits_url,
+          git_commits_url: repo.git_commits_url,
+          created_at: repo.created_at,
+          updated_at: repo.updated_at,
+          pushed_at: repo.pushed_at,
+          language: repo.language,
+          forks_count: repo.forks_count,
+          open_issues_count: repo.open_issues_count,
+          default_branch: repo.default_branch,
+          stargazers_count: repo.stargazers_count,
+        }));
+        setRepos(newRepos);
+        setRepoNumber(repos);
+        return newRepos;
+      } else {
+        console.error("Request failed with status:", res.status);
+        setRepos([]);
+        return [];
+      }
+    } catch (error) {
+      console.error("Error fetching repositories:", error);
+      setRepos([]);
+      return [];
+    }
+  }
+};
+
+export const getProfile = async (
+  profileName: string,
+  setProfile: {
+    (value: SetStateAction<ProfileItem | null>): void;
+  }
+) => {
+  if (profileName) {
+    const res = await octokit.request(
+      `GET https://api.github.com/users/${profileName}`
+    );
+
+    if (res.status === 200) {
+      const profile = {
+        name: res.data.name,
+        bio: res.data.bio,
+        login: res.data.login,
+        avatar_url: res.data.avatar_url,
+        followers: res.data.followers,
+        following: res.data.following,
+        organizations_url: res.data.organizations_url,
+      };
+      setProfile(profile);
+      return profile;
+    } else {
+      console.error("Request failed with status:", res.status);
+      return;
     }
   }
 };
