+++ src/userForm/UserForm.tsx
@@ -4,9 +4,9 @@ import { useNavigate } from "@tanstack/react-router";
 import { Field, Form, Formik } from "formik";
 import { useEffect, useState } from "react";
 import Select, { SingleValue } from "react-select";
-import { octokit } from "../../environment/apiKey";
 import { Option } from "../Models/interfaces";
 import Drawer from "../Navigation/Drawer";
+import { getUsers } from "./Api/userFormApi";
 
 function UserForm() {
   const navigate = useNavigate({ from: "/profile" });
@@ -16,7 +16,7 @@ function UserForm() {
     useState<SingleValue<Option> | null>(null);
 
   useEffect(() => {
-    getUsers();
+    getUsers(searchTerm, setOptions);
   }, []);
 
   //when the search term, or new options, have changed, check the search term
@@ -38,47 +38,17 @@ function UserForm() {
       );
 
       if (!searchTermContainsInOptions) {
-        getUsers();
+        getUsers(searchTerm, setOptions);
       }
     } else {
       setOptions([]);
     }
   }, [searchTerm, options]);
 
-  //fetch the users from the github api
-  //populate the options array with the fetched data
-  const getUsers = async () => {
-    if (searchTerm) {
-      try {
-        const res = await octokit.request(
-          `GET https://api.github.com/search/users?q=${searchTerm?.trim()}&per_page=100`
-        );
-        let data: any;
-        if (res.status === 200) {
-          data = res.data;
-        } else {
-          // Handle errors appropriately
-          console.error("Request failed with status:", res.status);
-        }
-        const newOptions =
-          data?.items?.map((user: { login: string }) => ({
-            value: user.login,
-            label: user.login,
-          })) ?? [];
-        setOptions(newOptions); // Set options to the newly fetched data
-        return newOptions;
-      } catch (error) {
-        console.error("Error fetching usernames:", error);
-        setOptions([]);
-        return [];
-      }
-    }
-  };
-
   //tanstack/react-query hook to fetch the users
   useQuery({
     queryKey: ["searchUsernames", searchTerm?.trim()],
-    queryFn: getUsers,
+    queryFn: () => getUsers(searchTerm, setOptions),
     enabled: false,
   });
